% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nyc311}
\alias{etl_extract.etl_nyc311}
\title{Using NYC311 data}
\usage{
\method{etl_extract}{etl_nyc311}(obj, years = lubridate::year(Sys.Date()),
  months = lubridate::month(Sys.Date()), n = 1e+06, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{years}{years that the issues have been created (the most recent year is the default)}

\item{months}{months that the issues have been created (the most recent month is the default)}

\item{n}{number of readings (1000000 is the default)}

\item{...}{arguments passed to \code{\link{download.file}}. Set method as "curl" for Linux system, and as default for Windows and Mac OS X systems.}
}
\description{
Using NYC311 data
}
\details{
This function downloads NYC311 data from NYC311 OPEN DATA website for years and months specified.
}
\examples{
\dontrun{
calls <- etl("nyc311", dir = "~/Desktop/nyc311")
calls \%>\%
  etl_extract(years = 2010:2011, months = 1:3, n =100) \%>\%
  etl_transform(years = 2010:2011, months = 1:3) \%>\%
  etl_load(years = 2010:2011, months = 1:3)

calls \%>\%
  tbl("calls") \%>\%
  glimpse()
  
calls <- calls \%>\%
  tbl("calls") \%>\%
  collect()
}
}

