% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nyc311}
\alias{etl_extract.etl_nyc311}
\alias{etl_transform.etl_nyc311}
\title{using 311 data}
\usage{
\method{etl_extract}{etl_nyc311}(obj, year = 1987:this_year, month = 1:12,
  day = 1:30, ...)

\method{etl_transform}{etl_nyc311}(obj, year = 2015, month = 3, day = 18,
  ...)
}
\arguments{
\item{year}{a numeric variable that inidicates the year}

\item{month}{a numeric variable that inidicates the month}

\item{day}{a numeric variable that inidicates the day}
}
\description{
using 311 data
}
\examples{

calls <- etl("nyc311", dir = "~/Desktop/nyc311")
calls \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()

}

