% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nyc311}
\alias{etl_extract.etl_nyc311}
\alias{etl_load.etl_nyc311}
\alias{etl_transform.etl_nyc311}
\title{Using NYC311 data}
\usage{
\method{etl_extract}{etl_nyc311}(obj, begin = Sys.Date() - 2,
  end = Sys.Date() - 1, n = 1e+06, ...)

\method{etl_transform}{etl_nyc311}(obj, begin = Sys.Date() - 2,
  end = Sys.Date() - 1, ...)

\method{etl_load}{etl_nyc311}(obj, schema = FALSE, begin = Sys.Date() - 2,
  end = Sys.Date() - 1, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{begin}{begin date}

\item{end}{end date}

\item{n}{number of readings (has to be less than or equal to one million)}

\item{...}{arguments passed to methods}

\item{schema}{Either a logical, a filename pointing to SQL schema file, or a character
vector containing the schema itself. If \code{schema = TRUE}, then the built-in
schema will be used. Note
that the flavor of SQL in this file must match the type of the source. That is,
if your object is of type \code{\link[dplyr]{src_mysql}}, then make sure that
the schema you specify here is written in MySQL (and not PostgreSQL). Please
note that SQL syntax is not, in general, completely portable. Use with caution, as this may
clobber any existing data you have in an existing database.}
}
\description{
Using NYC311 data
}
\examples{
calls <- etl("nyc311", dir = "~/Desktop/nyc311")
calls \%>\%
  etl_extract(n = 10) \%>\%
  etl_transform() \%>\%
  etl_load()

calls \%>\%
  tbl("calls") \%>\%
  glimpse()


}

