% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_nyc311}
\alias{etl_extract.etl_nyc311}
\alias{etl_transform.etl_nyc311}
\title{Using NYC311 data}
\usage{
\method{etl_extract}{etl_nyc311}(obj, begin = "2010-01-01":today_date,
  end = "2010-01-01":today_date, ...)

\method{etl_transform}{etl_nyc311}(obj, begin = "2010-01-01":today_date,
  end = "2010-01-01":today_date, ...)
}
\arguments{
\item{begin}{begin date}

\item{end}{end date}
}
\description{
Using NYC311 data
}
\examples{
calls <- etl("nyc311", dir = "~/Desktop/nyc311")
calls \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()

}

