% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_load.R, R/etl_transform.R
\name{etl_extract.etl_nyc311}
\alias{etl_extract.etl_nyc311}
\alias{etl_load.etl_nyc311}
\alias{etl_transform.etl_nyc311}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_nyc311}(obj, years = lubridate::year(Sys.Date()),
  months = lubridate::month(Sys.Date()), num_calls = 1e+06, ...)

\method{etl_load}{etl_nyc311}(obj, years = lubridate::year(Sys.Date()),
  months = lubridate::month(Sys.Date()), ...)

\method{etl_transform}{etl_nyc311}(obj, years = lubridate::year(Sys.Date()),
  months = lubridate::month(Sys.Date()), ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{years}{years that the issues have been created (the most recent year is the default)}

\item{months}{months that the issues have been created (the most recent month is the default)}

\item{num_calls}{number of readings (1000000 is the default)}

\item{...}{arguments passed to \code{\link{download.file}}. Set method as "curl" for Linux system, and as default for Windows and Mac OS X systems.}
}
\description{
etl_extract

etl_load

etl_transform
}
\details{
This function downloads NYC311 data from NYC311 OPEN DATA website for years and months specified.

This function loads NYC311 data into a local database for years and months specified.

This function transforms NYC311 data for years and months specified.
}
\examples{
\dontrun{
calls <- etl("nyc311")
calls \%>\%
  etl_init() \%>\%
  etl_update(years = 2010:2011, months = 1:3, num_calls = 100)

calls \%>\%
  tbl("calls") \%>\%
  glimpse()
  
calls_df <- calls \%>\%
  tbl("calls") \%>\%
  collect()
}
}

